--- src/components/features/LoginForm/Login.module.css
@@ -1,3 +1,0 @@
-.error{
-    margin-top: 8px;
-}
