+++ src/components/features/play/ExitButton/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import { DefaultButton } from "../../ui/Button";
-import { Modal } from "../../ui/Modal";
+import { DefaultButton } from "../../../ui/Button";
+import { Modal } from "../../../ui/Modal";
 import styles from "./index.module.css";
 import { useNavigate } from "react-router-dom";
