--- src/components/features/tab/tabs.css
@@ -1,64 +1,0 @@
-ul, li {
-    margin: 0;
-    padding: 0;
-  }
-  
-  .tabs {
-    margin: 0 auto;
-    width: 100%;
-  }
-  
-  .tab-btn {
-    display: flex;
-    width: 90%;
-    gap: 4px;
-  }
-  
-  .tab-name {
-    cursor: pointer;
-    display: block;
-    width: calc(60% );
-    padding: 10px 16px 6px;  
-    font-size: 16px;
-    font-weight: bold;
-    text-align: center;
-    transition: ease 0.3s;
-    border-radius: 50px 50px 0 0;
-    white-space: nowrap;
-    margin-left: 22px;
-  }
-
-  .tab-name-one {
-    background-color: var(--redorange);
-    color: var(--black);
-  }
-  
-  .tab-name-two {
-    background-color: var(--brown);
-    color: var(--white);
-  }
-  
-  .tab-name:hover {
-    opacity: 0.7;
-  }
-  
-  .tab-content {
-    width: 280px;
-    height: 300px;
-    padding: 20px;
-    font-size: 16px;
-    border-radius: 8px;
-  }
-  
-  .tab-content-one {
-    background-color: var(--redorange);
-  }
-  
-  .tab-content-two {
-    background-color: var(--brown);
-  }
-  
-  .tab-name.is-active {
-    opacity: 1;
-    transition: 0.3s ease;
-  }
