+++ src/components/features/welcome/LoginForm/index.module.css
@@ -1,0 +1,13 @@
+.login-form {
+  background-color: transparent;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+}
+
+.error {
+  margin-top: 8px;
+}
