+++ src/components/features/welcome/LoginForm/index.tsx
@@ -1,9 +1,9 @@
-import { PasswordInput } from "../../ui/PasswordInput";
-import { TextInput } from "../../ui/TextInput";
+import { PasswordInput } from "../../../ui/PasswordInput";
+import { TextInput } from "../../../ui/TextInput";
 import { WelcomeButton } from "../WelcomeButton";
 import { useState } from "react";
-import LoginAnonymous from "./LoginAnonimos";
-import styles from "./Login.module.css";
+import LoginAnonymous from "../LoginAnonimos";
+import styles from "./index.module.css";
 
 interface LoginProps {
 	onSubmit: (name: string, password: string) => void;
@@ -20,7 +20,7 @@ const Login: React.FC<LoginProps> = ({ onSubmit, loginError }) => {
 	};
 
 	return (
-		<form onSubmit={handleSubmit}>
+		<form className={styles["login-form"]} onSubmit={handleSubmit}>
 			<TextInput
 				label={"おなまえ"}
 				type="text"
@@ -32,8 +32,11 @@ const Login: React.FC<LoginProps> = ({ onSubmit, loginError }) => {
 				value={password}
 				onChange={(e) => setPassword(e.target.value)}
 			/>
-			<WelcomeButton color="brown">あそびにいく</WelcomeButton>
-			{loginError && <p className={styles.error}>{loginError}</p>}
+			<div>
+				<WelcomeButton color="brown">あそびにいく</WelcomeButton>
+				{loginError && <p className={styles.error}>{loginError}</p>}
+			</div>
+
 			<LoginAnonymous />
 		</form>
 	);
