+++ src/components/features/welcome/SignupForm/index.module.css
@@ -1,0 +1,9 @@
+.login-form {
+  background-color: transparent;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+}
