+++ src/index.css
@@ -15,9 +15,16 @@
   font-style: normal;
 }
 
+html,
 body {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
   margin: 0;
   padding: 0;
+  width: 100%;
+  height: 100%;
   box-sizing: border-box;
   color: #080000;
   background-color: var(--gray);
@@ -41,6 +48,7 @@ h4 {
 
 p {
   font-size: 12px;
+  margin: 0;
 }
 
 footer {
@@ -54,5 +62,11 @@ footer {
 
 a {
   font-size: 12px;
-  background-color: transparent;
+  background-color: var(--blue);
+}
+
+ul,
+li {
+  margin: 0;
+  padding: 0;
 }
