--- src/pages/Welcome/Welcome.tsx
@@ -1,47 +1,0 @@
-import type React from "react";
-import "./welcome.css";
-import Signup from "../../components/features/SignupForm/Signup";
-import Layout from "../../layout";
-import Login from "../../components/features/LoginForm/Login";
-import Tab from "../../components/features/tab/Tab";
-
-interface WelcomeProps {
-	onLoginSubmit: (name: string, password: string) => void;
-	onSignupSubmit: (name: string, password: string) => void;
-	loginError: string | null;
-}
-
-const Welcome: React.FC<WelcomeProps> = ({
-	onLoginSubmit,
-	onSignupSubmit,
-	loginError,
-}) => {
-	return (
-		<Layout>
-			<div className="welcome-body">
-				<h2>ばーちゃるぼっくす</h2>
-				<Tab
-					labels={["ろぐいん", "とうろく"]}
-					contents={[
-						<Login
-							key="login"
-							onSubmit={onLoginSubmit}
-							loginError={loginError}
-						/>,
-						<Signup key="signup" onSubmit={onSignupSubmit} />,
-					]}
-					id={""}
-				/>
-				<img
-					src="/cats/catsTower-circle.png"
-					className="main-logo"
-					alt="logo"
-					width="84"
-					height="132"
-				/>
-			</div>
-		</Layout>
-	);
-};
-
-export default Welcome;
