--- src/pages/Welcome/welcome.css
@@ -1,45 +1,0 @@
-
-body, html {
-    margin: 0;
-    padding: 0;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  .welcome-body {
-    width: 100%;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 30px;
-  }
-  
-  form {
-    background-color: transparent;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 16px;
-  }
-
-  .main-logo{
-    position: absolute;
-    background-color: transparent;
-    top: 72%;
-    right: 4%;
-  }
-  
-
-a {
-  background-color: transparent;
-  color: var(--blue);
-}
-
